Convert CMakeList.txt to support Other TLS libraries (#556)
* Convert CMakeList.txt to support Other TLS libraries

Currently, the CMake infrastructure for lsquic assumes that boring ssl
is the target tls library

Given the desire to be able to link to other TLS libraries, lets start
by updating CMakeLists to do the following:

1) Make the designation of the TLS library location generic.  Instead of
   calling it BORINGSSL_DIR, specify the Cmake variable as LIBSSL_DIR

2) Detect what kind of quic api is in use.  Currently, most major TLS
   libraries use one of two 3rd party quic stack apis:
        a) SSL_provide_quic_data
        b) SSL_set_quic_tls_cbs
   Augment the CMakeLists.txt to set a cflag and a cmake variable to
   record which api is being targeted.  This will let us, in future
   commits, identify which sources to compile to target the appropriate
   api.

Note, this is part of an effort to allow lsquic to support the canonical
OpenSSL sources.  OpenSSL can be targeted with this change, but lsquic
won't build with the latter (yet), this just paves the way to allow us
to try build with OpenSSL and identify which TLS/crypto apis need to be
updated to support a common set of calls so either library can be used
in the future.

* Update appveyor to use new LIBSSL_DIR macro

* update cirrus.yml to use new LIBSSL_DIR macro

* Update Dockerfile to use new LIBSSL_DIR macro

* Update Windows documentation to reflect new LIBSSL_DIR macro