Fix up sha256, lshkdf_* and lsquic_c255_* apis to be compatible with openssl (#557)
* Update target boringssl version

The boringssl version that lsquic compiles with varies in ci targets and
is quite old.  Specifically it is lacking some api support that allows
for reasonable compatibility with current versions of openssl.  Update
to the latest release to complete compatibility.

* modify lshkdf_expand to use openssl compatible api

the lshkdf_expand function use an hkdf api that is unique to boringssl.
In order to make it amenable to using openssl at well, we need to
convert it to use apis that exist in both implementations

* convert sha256 function to use compatible openssl api

The sha256 function uses a api that is unique to boringssl.  Convert it
to an api that exists in both boringssl and openssl so that either
implementation can be used

* use compatible api for x25519 key deriviation

The x25519 api in lsquic_crypto.c uses a custom boringssl api.  Update
it to use a compatible api that works in both boring and openssl

* update appveyor to use windows visual studio 2019

When updating to a more recent boringssl to pull in more recent api
updates, the boringssl build broke, as visual studio 2017 doesn't
support the _udiv128 intrinsic that boringssl requires.  We should
update to at least VS 2019 images as:
1) VS 2017 is going EOL in 2027
2) Not updating prevents pulling in newer boringssl versions that fix
   security issues

* Update check_symbol_exists for older compilers

It appears older compilers (like on ubuntu 20.04) don't resolve symbols
properly when -l options are specified before static libraries, leading
to failed symbols resolution and configuration errors

Update the check_symbol_exists code in CMakeLists.txt to ensure that
-pthread and -lstdc++ get added to the end of the command line

* Fix windows build with new boringssl

Windows builds are failing with the updated boringssl as boring uses
some public structures with unnamed structs, which Visual C warns about.
As warnings are treated as errors, this fails the build.

Easy fix, just suppress the warnings about unnamed structs embedded
within other structs (W4201), as the language and compiler actually
handle it just fine.

* Update appveyor to use new cmake command syntax

with the use of the VS 2019 image, the new version of cmake no longer
supports the --config option, we need to use -DCMAKE_BUILD_TYPE