Allow split buffered packet to generate one packet
This appears to be a case where DPLPMTUD increases a path's MTU.  This
makes it possible for a packet that needs to be split to generate just
one packet, as the newly allocated packet is larger than the original
packet.

It is possible to optimize this scenario, but it does not happen often
and so it is better to keep the code simple.

Fixes issue #505.